+++ src/slides.js
@@ -1471,6 +1471,40 @@ const slides = [
       mediaContain: true,
     },
   },
+  {
+    id: "slide25-instagram-no-max-entries",
+    type: "instagram-feed",
+    darkModeEnabled: false,
+    content: {
+      duration: 5000,
+      hashtagText: "#myhashtag",
+      orientation: "landscape",
+      imageWidth: 40.0,
+      entryDuration: 10,
+      maxEntries: "",
+      mediaContain: true,
+    },
+    feedData: [
+      {
+        text: "#mountains #horizon",
+        textMarkup:
+          '<div class="text">Sed nulla lorem, varius sodales justo ac, ultrices placerat nunc.</div>\n<div class="tags"><span class="tag">#mountains</span> <span class="tag">#horizon</span> Lorem ipsum ...</div>',
+        mediaUrl: "./fixtures/images/mountain1.jpeg",
+        videoUrl: null,
+        username: "username",
+        createdTime: "2022-02-03T08:50:07",
+      },
+      {
+        text: "#mountains #horizon #sky",
+        textMarkup:
+          '<div class="text">Aenean consequat sem ut tortor auctor, eget volutpat libero consequat. Donec lacinia varius quam, ut efficitur diam ultrices et. Aliquam eget augue at felis rhoncus egestas. Sed porttitor elit a tellus tempus, sed tempus sapien finibus. Nam at dapibus sem. Aliquam sit amet feugiat ex. Ut dapibus, mi eu fermentum dignissim, sem ipsum vulputate est, sit amet euismod orci odio pharetra massa.</div>\n<div class="tags"><span class="tag">#mountains</span> <span class="tag">#horizon</span> <span class="tag">#sky</span> Lorem ipsum and mountains ...</div>',
+        mediaUrl: "./fixtures/images/mountain2.jpeg",
+        videoUrl: null,
+        username: "username2",
+        createdTime: "2022-01-03T08:50:07",
+      },
+    ],
+  },
 ];
 
 export default slides;